Update config.json.example to include page_size_limit
page_size_limit is supposed to be user configurable in order to account for hundreds of rows with the exact same updated timestamp. In Stitch, this is not available in the GUI and the integration runs with a fixed page_size_limit value of 100, causing pagination issues.
I don't assume this example file will directly affect the Stitch GUI, but the example file should probably include it at least.