Update sync.py to increase RESULT_RETURN_LIMIT
As per comments in the code and first hand experience with eternal pagination as outlined in issue #12 , RESULT_RETURN_LIMIT needs to be increased for to account for hundreds of rows with the same updated_at timestamp. API docs say the max value for page size is 500, and testing done by @mascah proved that even setting a page size value of 1000 only returned 500 rows.